desighn pattern iterator
